refactor: code refactor backend implementation
