+++ hlskit-rs/src/backends/mod.rs
@@ -38,7 +38,5 @@
  * The use of the unmodified library in proprietary software is governed solely by the LGPLv3.
  */
 
-pub mod ffmpeg_command_builder;
-pub mod hlskit_error;
-pub mod m3u8_tools;
-pub mod segment_tools;
+pub mod ffmpeg_backend;
+pub mod gstreamer_backend;
