+++ hlskit-rs/src/tools/segment_tools.rs
@@ -38,4 +38,49 @@
  * The use of the unmodified library in proprietary software is governed solely by the LGPLv3.
  */
 
-pub mod hls_video_processing_service;
+use std::{fs::File, io::Read, path::PathBuf};
+
+use crate::{
+    models::hls_video::{HlsVideoResolution, HlsVideoSegment},
+    tools::hlskit_error::HlsKitError,
+};
+
+pub fn read_playlist_and_segments(
+    playlist_filename: &str,
+    segment_filename: &str,
+    resolution: (i32, i32),
+    stream_index: i32,
+) -> Result<HlsVideoResolution, HlsKitError> {
+    let mut resolution = HlsVideoResolution {
+        resolution,
+        playlist_name: format!("playlist_{stream_index}.m3u8"),
+        playlist_data: Vec::new(),
+        segments: Vec::new(),
+    };
+
+    // Read the playlist file
+    let mut playlist_file = File::open(playlist_filename)?;
+    playlist_file.read_to_end(&mut resolution.playlist_data)?;
+
+    // Read all segment files
+    let mut segment_index = 0;
+    loop {
+        let segment_path = segment_filename.replace("%03d", &format!("{segment_index:03}"));
+        if !PathBuf::from(&segment_path).exists() {
+            break;
+        }
+
+        let mut segment_file = File::open(&segment_path)?;
+        let mut segment_data = Vec::new();
+        segment_file.read_to_end(&mut segment_data)?;
+
+        let segment = HlsVideoSegment {
+            segment_name: format!("data_{stream_index}_{segment_index:03}.ts"),
+            segment_data,
+        };
+        resolution.segments.push(segment);
+        segment_index += 1;
+    }
+
+    Ok(resolution)
+}
